ai_backend/requirements.txt
fastapi
uvicorn
langchain-groq
langchain-core
langgraph
python-dotenv
